feat(horizontal line)!: change cursor end position to rest upon the final line segment drawn
Change ttui::draw::horizontal_line to now place the cursor upon the final line segment drawn at the conclusion of drawing the line in order to make drawing shapes more intuitive for developers.

BREAKING CHANGE: previously, cursor would land one column beyond the end of the line (1 column to the left if the line was drawn to the left, 1 column to the right if line was drawn to the right).  Now cursor lands on top of the final segment of the line drawn.